Split reconcile file
